requirements.txt
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
filelock==3.16.1
Flask==3.0.3
fsspec==2024.12.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20240930
packaging==24.2
pycparser==2.22
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sounddevice==0.5.1
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2